feat: imperfect stringer
this one is long overdue. this implementation is lacking in many department.
First, it doesn't take into consideration the type of the struct's field.
Second, it doesnt work with any type other than struct.
Third, I need to implement goroutine to make it faster.
Fourth, the module name should be changed to 'stringer' instead.
Fifth, the error handling is not well done.